requirements.txt
discord.py==2.3.2
transformers==4.41.2
datasets==2.19.2
accelerate==0.30.1
# You will also need the correct pytorch depending if u have a gpu: https://pytorch.org/get-started/locally/
